+++ builtInFunctions/creator_test.go
@@ -128,7 +128,7 @@ func TestCreateBuiltInContainter_Create(t *testing.T) {
 
 	err := f.CreateBuiltInFunctionContainer()
 	assert.Nil(t, err)
-	assert.Equal(t, f.BuiltInFunctionContainer().Len(), 33)
+	assert.Equal(t, f.BuiltInFunctionContainer().Len(), 32)
 
 	err = f.SetPayableHandler(nil)
 	assert.NotNil(t, err)
