+++ mock/enableEpochsHandlerStub.go
@@ -34,7 +34,7 @@ type EnableEpochsHandlerStub struct {
 	IsMaxBlockchainHookCountersFlagEnabledField          bool
 	IsWipeSingleNFTLiquidityDecreaseEnabledField         bool
 	IsAlwaysSaveTokenMetaDataEnabledField                bool
-	IsMigrateUserNameEnabledEpochField                   bool
+	IsChangeUsernameEnabledEpochField                    bool
 	MultiESDTTransferAsyncCallBackEnableEpochField       uint32
 	FixOOGReturnCodeEnableEpochField                     uint32
 	RemoveNonUpdatedStorageEnableEpochField              uint32
@@ -259,9 +259,9 @@ func (stub *EnableEpochsHandlerStub) IsAlwaysSaveTokenMetaDataEnabled() bool {
 	return stub.IsAlwaysSaveTokenMetaDataEnabledField
 }
 
-// IsMigrateUsernameEnabled -
-func (stub *EnableEpochsHandlerStub) IsMigrateUsernameEnabled() bool {
-	return stub.IsMigrateUserNameEnabledEpochField
+// IsChangeUsernameEnabled -
+func (stub *EnableEpochsHandlerStub) IsChangeUsernameEnabled() bool {
+	return stub.IsChangeUsernameEnabledEpochField
 }
 
 // IsInterfaceNil -
